Fix DTO contract mismatches found by integration testing
Request DTOs:
- ShipmentRequest: send dimensions as dimension1/2/3 (cm) — the API
  silently ignores the length/width/height keys in requests, so orders
  were created with default dimensions and undervalued prices
- ShipmentRequest: weight is float (kg) to support sub-kilogram parcels
- CodRequest: add bankaccount / bankaccount_id payout fields
- Address: add state_code (required e.g. for US) and is_residential

Response DTOs:
- Order/OrderShort/ValuationPrice: use Service::tryFrom() so unknown
  service ids no longer throw ValueError; expose raw id via
  getServiceId(), getService() is now nullable
- Order: read the order-level VAT key as serialized by the API
  (price_var, with price_vat fallback); it carries the VAT rate
- Order: pickup is an object in real payloads — map it to a new
  OrderPickup DTO instead of ?string (TypeError on live data)
- Order/OrderShort: externalId is serialized as boolean false when
  absent (CBL-only field) — normalize to null (TypeError on live data)
- Shipment: weight/weight_billable are float
- Valuation: keep price table entries for service ids missing from the
  enum instead of crashing

Enum:
- add services 32/33 (DPD Pickup to parcel machines) and 71
  (Poczta Polska Ukraina) — all three are live on production

+++ src/DTO/Response/Shipment.php
@@ -7,8 +7,9 @@ namespace AlsendoOne\SDK\DTO\Response;
 class Shipment
 {
     private string $shipmentTypeCode;
-    private int $weight;
-    private int $weightBillable;
+    /** Weight in kilograms. */
+    private float $weight;
+    private float $weightBillable;
     private int $length;
     private int $width;
     private int $height;
@@ -22,8 +23,8 @@ class Shipment
 
     private function __construct(
         string $shipmentTypeCode,
-        int $weight,
-        int $weightBillable,
+        float $weight,
+        float $weightBillable,
         int $length,
         int $width,
         int $height,
@@ -57,8 +58,8 @@ class Shipment
     {
         return new self(
             (string) ($data['shipment_type_code'] ?? ''),
-            (int) ($data['weight'] ?? 0),
-            (int) ($data['weight_billable'] ?? 0),
+            (float) ($data['weight'] ?? 0),
+            (float) ($data['weight_billable'] ?? 0),
             (int) ($data['length'] ?? 0),
             (int) ($data['width'] ?? 0),
             (int) ($data['height'] ?? 0),
@@ -77,12 +78,12 @@ class Shipment
         return $this->shipmentTypeCode;
     }
 
-    public function getWeight(): int
+    public function getWeight(): float
     {
         return $this->weight;
     }
 
-    public function getWeightBillable(): int
+    public function getWeightBillable(): float
     {
         return $this->weightBillable;
     }
